backend/requirements.txt
fastapi 
uvicorn 
pillow 
google-generativeai 
requests 
python-multipart
python-dotenv
openai
whisper 
ffmpeg-python